SEQNG-724: Extracted methods from GPIClient to GiapiClient that will be used by all clients.
